chore: <Alarm> -> <AlarmBell>로 컴포넌트 이름 변경 (SP-574)
