+++ src/Components/Header/index.tsx
@@ -12,7 +12,7 @@ import Dropdown from '../Dropdown';
 import { useModalStore } from '@/store/modal';
 import { useLogOutMutation } from '@/Hooks/auth/useLogOutMutation';
 import DropdownItem from '../Common/DropdownItem';
-import { AlarmBell } from './Alarm';
+import { AlarmBell } from './AlarmBell';
 
 /** 사이트 메인 헤더 */
 const Header = () => {
